Prevent ball flashing
The ball will flashing because we didn't apply double buffer.
And that's impossible in MARS Bitmap Display.
So We doing something tricky to prevent flashing.

Once we need to redraw the ball, fisrt we cover it with a color which
very similar to ball's color(ballPostColor).

Next we draw the ball on its next position.

And then clear any ballPostColor remain in the ball's original position.

This strategy will reduce almost 80% of flashing, but not 100% since the
display is so slow, the process might be witness by human eye.